+++ src/django_idom/__init__.py
@@ -1,7 +1,14 @@
-from . import components, hooks
-from .websocket.consumer import IdomWebsocket
-from .websocket.paths import IDOM_WEBSOCKET_PATH
+from django_idom import components, decorators, hooks, types
+from django_idom.types import IdomWebsocket
+from django_idom.websocket.paths import IDOM_WEBSOCKET_PATH
 
 
 __version__ = "1.1.0"
-__all__ = ["IDOM_WEBSOCKET_PATH", "IdomWebsocket", "hooks", "components"]
+__all__ = [
+    "IDOM_WEBSOCKET_PATH",
+    "types",
+    "IdomWebsocket",
+    "hooks",
+    "components",
+    "decorators",
+]